[WFLY-20278] Remove the EJBs from the SSO testing, these had been disabled 11 years ago.
Although we still had EJBServlet it was just pretending to call the EJB.
