--- testsuite/shared/src/main/java/org/jboss/as/test/integration/web/sso/interfaces/StatelessSessionLocal.java
@@ -1,24 +1,0 @@
-/*
- * Copyright The WildFly Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-package org.jboss.as.test.integration.web.sso.interfaces;
-
-import java.rmi.RemoteException;
-
-import jakarta.ejb.EJBLocalObject;
-
-/**
- * A trivial SessionBean local interface.
- *
- * @author Scott.Stark@jboss.org
- */
-public interface StatelessSessionLocal extends EJBLocalObject {
-
-    /** A method that returns its arg */
-    String echo(String arg) throws RemoteException;
-
-    /** A method that does nothing. It is used to test call optimization. */
-    void noop();
-
-}
