+++ testsuite/shared/src/main/resources/org/jboss/as/test/integration/web/sso/resources/application.xml
@@ -25,7 +25,4 @@
             <context-root>/war6</context-root>
         </web>
     </module>
-    <module>
-        <ejb>jbosstest-web-ejbs.jar</ejb>
-    </module>
 </application>
